started adding api route to save to kv store but now have auth error
