+++ src/routes/+page.svelte
@@ -51,9 +51,13 @@
 				}, 2000);
 			}, 1000);
 			// post to an endpoint to record the results
-			const postURL = `https://www.canyoutoast.com/${effort.power.toFixed(0)}/${effort.timeS}/${api.client_id}`
-			console.log(postURL)
-			fetch(postURL, {method: "post"}).then(() => {console.log("fufilled")})
+			const response = fetch('/efford', {
+			method: 'POST',
+			body: JSON.stringify(effort),
+			headers: {
+				'Content-Type': 'application/json'
+			}
+		}).then(() => console.log("fufilled"))
 		}).catch((error: Error) => {
 			if (svelteBrowser) {
 				// delete cookie
